重构：Pet，Pet继承了Item，却又作为Item属性使用。直接使用Pet，PetId通过YitIdHelper生成

+++ src/Application.Core/Game/Players/Pet.cs
@@ -39,7 +39,7 @@ namespace Application.Core.Game.Players
             }
         }
 
-        public sbyte getPetIndex(int petId)
+        public sbyte getPetIndex(long petId)
         {
             Monitor.Enter(petLock);
             try
